Print a space after printQualType
This fixes printing of Dvars of type Tbool.
Before this patch they are printed as `(Dvar "_ZN2Hp14_iommu_enabledE"
TboolNone)`, which leads to "Error: The reference TboolNone was not found in the current environment".
This patch changes the output to
`(Dvar "_ZN2Hp14_iommu_enabledE" Tbool None)`, which compiles correctly.