EXPERIMENT: Try out [Loc := ptr] and [Offset := offset].
Very large changeset that changes the underlying definitions of [Loc] and
[Offset] to be the primitives defined in the new pointer model.
It also converts to using the notations from [heap_notations.v] pervasively
(this seemed to ease porting since the argument order of [_offset_ptr] is
different from [_offsetL]).

Generally, the changes are mostly cosmetic. We avoid a lot of existential
quantifiers and all uses of [addr_of]. These were pretty painful to work with
in the original semantics, so there is likely a lot of benefit here.

There are still a few things that are stubbed out, e.g. [_local], [_this], and
[_result], and a few definitions that should be moved.