Adding deBruijn levels to the syntax + fix OpaqueValueExpr
This allows us to disambiguate the different binders for different
loops. Note that we assume in clang that the index always refers to the
innermost loop. It would be easy to change this, but I'm not sure how to
construct better information.

This also fixes the implementation of OpaqueValueExpr which
represents anonymous variables/variable references.