Manually mangle destructor names.
In certain situations (see comments) we might have a class which we have
not generated a destructor for which means that we can not ask clang to
mangle the name of the destructor. Therefore, we manually mangle all
destructor names.

We only need to do this for destructors because destructors are the only
name that we *must* have.