Add missing Proper instances for `elements`, `list_bind`, `list_to_set`
These are needed to prove monotonicity for `set_concat_bind` and
`finite_preimage_set`.

Note: those are already provable directly via `set_solver`, but that does not
help e.g. the `rewrite` tactic.

Also add a Params instance for for `list_to_set`; it's already there for `mbind`
and `elements`.