BUG: Introducing [Eoperator_call] and order-of-evaluation fixes.
This introduces [Eoperator_call] which is either a function call
or a member-function call but the order of evaluation of the expression
is dictated by the order of evaluation of the operator being called.

To support this, we lift out [eval] which encapsulates different
evaluation strategies.

A side effect of this change is that it requires us to perform virtual
function resolution immediately before the call occurs rather than
immediately after the function argument is evaluated.